Add the error handling for cascade deletion error and also add tests

+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
@@ -44,10 +48,21 @@ export class UsersService {
   }
 
   async remove(id: number) {
-    const user = await this.repo.findOneBy({ id });
+    const user = await this.repo.findOne({
+      where: { id },
+      relations: ['reports'],
+    });
+
     if (!user) {
       throw new NotFoundException('user not found');
     }
-    return this.repo.remove(user);
+
+    if (user.reports.length > 0) {
+      throw new BadRequestException(
+        'User has reports associated with their account. Please delete all reports before deleting the user.',
+      );
+    }
+
+    return await this.repo.remove(user);
   }
 }
